Scenario two complete Proceed to checkout is side function
